+++ client/App.js
@@ -1,15 +1,15 @@
-import React from 'react'
+import React from 'react';
 
-import Navbar from './components/Navbar'
-import Routes from './Routes'
+import Navbar from './components/Navbar';
+import AppRoutes from './AppRoutes';
 
 const App = () => {
   return (
     <div>
       <Navbar />
-      <Routes />
+      <AppRoutes />
     </div>
-  )
-}
+  );
+};
 
-export default App
+export default App;