+++ client/AppRoutes.js
@@ -1,0 +1,47 @@
+import React, { useEffect } from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { Route, Routes } from 'react-router-dom';
+import AuthForm from './components/AuthForm';
+import Home from './components/Home';
+import { me } from './store';
+
+/**
+ * COMPONENT
+ */
+
+const AppRoutes = () => {
+  const isLoggedIn = useSelector((state) => !!state.auth.id);
+  const dispatch = useDispatch();
+
+  useEffect(() => {
+    dispatch(me());
+  }, []);
+
+  return (
+    <div>
+      {isLoggedIn ? (
+        <Routes>
+          <Route path="/*" element={<Home />} />
+          <Route to="/home" element={<Home />} />
+        </Routes>
+      ) : (
+        <Routes>
+          <Route
+            path="/*"
+            element={<AuthForm name="login" displayName="Login" />}
+          />
+          <Route
+            path="/login"
+            element={<AuthForm name="login" displayName="Login" />}
+          />
+          <Route
+            path="/signup"
+            element={<AuthForm name="signup" displayName="Sign Up" />}
+          />
+        </Routes>
+      )}
+    </div>
+  );
+};
+
+export default AppRoutes;