+++ client/components/Navbar.js
@@ -1,11 +1,16 @@
-import React from 'react'
-import { useSelector, useDispatch } from 'react-redux'
-import {Link} from 'react-router-dom'
-import {logout} from '../store'
+import React from 'react';
+import { useSelector, useDispatch } from 'react-redux';
+import { Link, useNavigate } from 'react-router-dom';
+import { logout } from '../store';
 
-const Navbar = () =>  {
-  const isLoggedIn = useSelector(state => !!state.auth.id)
-  const dispatch = useDispatch()
+const Navbar = () => {
+  const isLoggedIn = useSelector((state) => !!state.auth.id);
+  const dispatch = useDispatch();
+  const navigate = useNavigate();
+  const logoutAndRedirectHome = () => {
+    dispatch(logout());
+    navigate('/login');
+  };
 
   return (
     <div>
@@ -15,9 +20,9 @@ const Navbar = () =>  {
           <div>
             {/* The navbar will show these links after you log in */}
             <Link to="/home">Home</Link>
-            <a href="#" onClick={ () => dispatch(logout()) }>
+            <button type="button" onClick={logoutAndRedirectHome}>
               Logout
-            </a>
+            </button>
           </div>
         ) : (
           <div>
@@ -29,7 +34,7 @@ const Navbar = () =>  {
       </nav>
       <hr />
     </div>
-    )
-}
+  );
+};
 
-export default Navbar
+export default Navbar;