+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 ### Dependencies
 
+- _remove_ history (no longer used with react-router-dom version 6)
+
 ### DevDependencies
 
 - @babel/core update to 7.19.6 from 7.12.10
@@ -11,6 +13,7 @@
 - @babel/preset-react update to 7.18.6 from 7.12.10
 - @babel/register update to 7.18.9 from 7.12.10
 - babel-loader update to 8.2.5 from 8.2.2
+- react-router-dom update to 6.4.2 from 5.0.0
 - webpack update to 5.74.0 from 5.15.0
 - webpack-cli update to 4.10.0 from 4.3.1
 - _remove_ enzyme (deprecated)
@@ -18,6 +21,7 @@
 
 ### Refactor
 
+- update components and store for compatiblity with React Router version 6
 - update webpack.config.js to follow best practices & include babel/preset-env
 - remove .babelrc (babel configuration now in webpack.config.js)
 - remove Home component testing until full testing tech stack change can be prioritized (enzyme deprecated)